+++ resolwe_bio/tests/processes/test_diff_expression.py
@@ -1,3 +1,5 @@
+from pathlib import Path
+
 from resolwe.flow.models import Data, Process
 from resolwe.test import tag_process, with_resolwe_host
 
@@ -121,6 +123,86 @@ class DiffExpProcessorTestCase(KBBioProcessTestCase):
             )
         ]
         self.assertEqual(deseq2.process_error, error_msg)
+
+    @with_resolwe_host
+    @tag_process("differentialexpression-deseq2")
+    def test_deseq2_nanostring(self):
+        base = Path("test_nanostring_deseq2")
+        inputs = base / "inputs"
+        outputs = base / "outputs"
+        with self.preparation_stage():
+
+            expression = Process.objects.create(
+                name="Upload nanostring expression data mock process",
+                requirements={
+                    "expression-engine": "jinja",
+                    "resources": {"network": True,},
+                    "executor": {"docker": {"image": "resolwebio/base:ubuntu-18.04",},},
+                },
+                contributor=self.contributor,
+                type="data:expression:nanostring:",
+                entity_type="sample",
+                entity_descriptor_schema="sample",
+                data_name="{{ src.file }}",
+                input_schema=[{"name": "src", "type": "basic:file:",},],
+                output_schema=[
+                    {"name": "exp", "type": "basic:file:",},
+                    {"name": "species", "type": "basic:string:",},
+                    {"name": "build", "type": "basic:string:",},
+                    {"name": "source", "type": "basic:string:",},
+                    {"name": "feature_type", "type": "basic:string:",},
+                ],
+                run={
+                    "language": "bash",
+                    "program": r"""
+re-import {{ src.file_temp|default(src.file) }} {{ src.file }} "txt" "txt" 0.1 compress
+re-save-file exp "${NAME}".txt.gz
+re-save species "Dictyostelium discoideum"
+re-save build "dd-05-2009"
+re-save source "DICTYBASE"
+re-save feature_type "gene"
+""",
+                },
+            )
+
+            exp_1 = self.run_process(
+                expression.slug, {"src": str(inputs / "exp_1_norm.txt.gz")},
+            )
+            exp_2 = self.run_process(
+                expression.slug, {"src": str(inputs / "exp_2_norm.txt.gz")},
+            )
+            exp_3 = self.run_process(
+                expression.slug, {"src": str(inputs / "exp_3_norm.txt.gz")},
+            )
+            exp_4 = self.run_process(
+                expression.slug, {"src": str(inputs / "exp_4_norm.txt.gz")},
+            )
+
+        inputs = {
+            "case": [exp_1.id, exp_3.id],
+            "control": [exp_2.id, exp_4.id],
+            "filter": {"min_count_sum": 0,},
+        }
+
+        diff_exp = self.run_process("differentialexpression-deseq2", inputs)
+
+        self.assertFileExists(diff_exp, "raw")
+        self.assertFile(
+            diff_exp,
+            "count_matrix",
+            str(outputs / "count_matrix.tab.gz"),
+            compression="gzip",
+        )
+        self.assertJSON(
+            diff_exp,
+            diff_exp.output["de_json"],
+            "",
+            str(outputs / "de_data_deseq.json.gz"),
+        )
+        self.assertFields(diff_exp, "source", "DICTYBASE")
+        self.assertFields(diff_exp, "species", "Dictyostelium discoideum")
+        self.assertFields(diff_exp, "build", "dd-05-2009")
+        self.assertFields(diff_exp, "feature_type", "gene")
 
     @with_resolwe_host
     @tag_process("differentialexpression-edger")